% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{calc_unemployment_rate}
\alias{calc_unemployment_rate}
\title{Make unemployment metric}
\usage{
calc_unemployment_rate(.data, var, ..., na.rm = TRUE)
}
\arguments{
\item{.data}{tibble}

\item{var}{unemployment variable}

\item{...}{grouping variables}
}
\value{
tibble
}
\description{
Make unemployment metric
}
\examples{
\dontrun{
cps \%>\% calc_unemployment_rate(employmentstatus)
acs \%>\% calc_unemployment_rate(employmentstatus)
}
}
