% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{make_education}
\alias{make_education}
\title{Convert education data to title case}
\usage{
make_education(.data, var)
}
\arguments{
\item{var}{}
}
\value{

}
\description{
Convert education data to title case
}
