% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{make_race_ethnicity}
\alias{make_race_ethnicity}
\title{Convert race / ethnicity to standard set}
\usage{
make_race_ethnicity(.data, var)
}
\arguments{
\item{.data}{tibble}

\item{var}{race variable}
}
\value{
tibble
}
\description{
Convert race / ethnicity to standard set
}
